+++ backend/src/samplespace/schemas/audio.py
@@ -5,5 +5,5 @@ from pydantic import BaseModel
 
 class AudioMetadata(BaseModel):
     key: str | None = None
-    bpm: float | None = None
+    bpm: int | None = None
     duration: float | None = None
